gitignore

+++ include/utils.h
@@ -6,14 +6,14 @@
 #include <stdlib.h>
 
 typedef struct node {
-    uint32_t val;
+    long double val;
     uint32_t count;
     struct node *next;
 } node_t;
 
-void push_end(node_t *head, uint32_t val);
-void push_head(node_t *head, uint32_t val);
-void ordered_insert(node_t *head, uint32_t val);
+void push_end(node_t *head, long double val);
+void push_head(node_t *head, long double val);
+void ordered_insert(node_t *head, long double val);
 void print_list(node_t *head);
 
 #endif
